google language detection fixed string tagger no entities in quotes
git-svn-id: https://141.76.40.86/svn-students/iircommon/toolkit/trunk@1205 5654db93-19e5-4567-beb8-64e8c25c7c45

Former-commit-id: 25eec504f892e7ff7d906c11159ff80603c28c7e

+++ src/main/java/tud/iir/helper/MathHelper.java
@@ -303,9 +303,9 @@ public class MathHelper {
      * @param number The integer to be transformed.
      * @return The IP address.
      */
-    public static String numberToIp(int number) {
+    public static String numberToIp(long number) {
         return (number >> 24 & 0xFF) + "." + (number >> 16 & 0xFF) + "." + (number >> 8 & 0xFF) + "."
-                + (number & 0xFF);
+        + (number & 0xFF);
     }
 
     /**
@@ -353,4 +353,5 @@ public class MathHelper {
 
         return alphaBeta;
     }
+
 }
